% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_hosp_status.R
\name{add_hosp_status}
\alias{add_hosp_status}
\title{Adds hospitalisation status to a data frame obtained using
\code{covidHubUtils::load_truth}}
\usage{
add_hosp_status(x)
}
\arguments{
\item{x}{A data frame to add hospitalisation status to}
}
\value{
an updated data frame
}
\description{
Adds hospitalisation status to a data frame obtained using
\code{covidHubUtils::load_truth}
}
\author{
Sebastian Funk
}
