% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_owid_to_weekly.R
\name{convert_owid_to_weekly}
\alias{convert_owid_to_weekly}
\title{Convert OWID truth data from daily to weekly}
\usage{
convert_owid_to_weekly(x, ...)
}
\arguments{
\item{x}{data to convert}

\item{...}{any parameters for \code{\link[=date_to_week_end]{date_to_week_end()}}}
}
\value{
converted data frame
}
\description{
Convert OWID truth data from daily to weekly
}
\author{
Sebastian Funk
}
