% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ensemble_relative_skill.R
\name{create_ensemble_relative_skill}
\alias{create_ensemble_relative_skill}
\title{Ensemble by relative skill}
\usage{
create_ensemble_relative_skill(
  forecasts,
  evaluation_date,
  continuous_weeks = 4,
  average = "mean",
  skill = "wis",
  history = "All",
  by_horizon = FALSE,
  eval_dir = here::here("evaluation", "weekly-summary"),
  return_criteria = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{forecasts}{\code{data.frame} containing all the forecasts to be summarised
as an ensemble.}

\item{evaluation_date}{which date should be used to measure model
performance? A corresponding file containing evaluations is expected in \code{evaluation/evaluation-{evaluation_date}.csv}}

\item{continuous_weeks}{include only forecasts with a history of evaluation}

\item{average}{Method to average. One of \code{"mean"} or \code{"median"}}

\item{skill}{the relative skill score to be used for creating the ensemble; a column called \code{rel_{skill}} is expected to exist in the evaluation csv file}

\item{history}{the number of recent weeks of history to consider in determining the weights; either "All" or a number of weeks}

\item{by_horizon}{whether to create the ensemble separately for each horizon (default: FALSE)}

\item{eval_dir}{character: the path in which to look for evaluation csv files}

\item{return_criteria}{logical : whether to return a model/inclusion criteria
grid as well as the ensemble forecast (default \code{TRUE})}

\item{verbose}{Logical determining whether diagnostic messages should
be printed while running (defaults to \code{FALSE}).}
}
\description{
Ensemble by relative skill
}
\details{
Ensemble = sum of forecast values weighted by the inverse of relative skill
Weights are by model, horizon, target, location
i.e. not weighted by quantile
}
