% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_preview.R
\name{link_preview}
\alias{link_preview}
\title{Generate link to the preview shiny app for forecast submissions}
\usage{
link_preview(gh_repo, pr_number, data_folder = "data-processed")
}
\arguments{
\item{gh_repo}{GitHub repository address in the format \code{username/repo}}

\item{pr_number}{Number of the pull request to validate}

\item{data_folder}{Path to the folder containing forecasts in this \code{gh_repo}}
}
\value{
A character vector containing one link for each of the forecast file
uploading in the PR #\code{pr_number}.
}
\description{
Generate link to the preview shiny app for forecast submissions
}
\examples{
link_preview(
  "epiforecasts/covid19-forecast-hub-europe",
  956
)

}
