% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_forecasts.R
\name{score_forecasts}
\alias{score_forecasts}
\title{Score forecasts}
\usage{
score_forecasts(forecasts, quantiles = NULL)
}
\arguments{
\item{forecasts}{\code{data.frame} containing all the forecasts to be summarised
as an ensemble.}

\item{quantiles}{Numeric vector of quantiles to be considered in the scoring.}
}
\description{
Score forecasts
}
