% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_models.R
\name{score_models}
\alias{score_models}
\title{Score models}
\usage{
score_models(forecasts, report_date, restrict_weeks = 0L, quantiles)
}
\arguments{
\item{forecasts}{\code{data.frame} containing all the forecasts to be summarised
as an ensemble.}

\item{report_date}{Date at which the scoring takes place}

\item{restrict_weeks}{Integer number of continuous weeks continuous weeks
leading up to the \code{report_date} the forecasts need to include to be
considered in the scoring.}

\item{quantiles}{Numeric vector of quantiles the forecasts need to include
to be considered in the scoring.}
}
\description{
Score models
}
