% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_scores.R
\name{summarise_scores}
\alias{summarise_scores}
\title{Summarise scores}
\usage{
summarise_scores(scores, report_date, restrict_weeks = 0L)
}
\arguments{
\item{scores}{table of scores, e.g. as created with \code{\link{score_models}}}

\item{report_date}{Date at which the scoring takes place}

\item{restrict_weeks}{Integer number of continuous weeks continuous weeks
leading up to the \code{report_date} the forecasts need to include to be
considered in the scoring.}
}
\description{
Summarise scores
}
