% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{date_to_week_end}
\alias{date_to_week_end}
\title{Convert a date to the last date of the week it's in}
\usage{
date_to_week_end(
  x,
  type = c("target", "forecast"),
  config_file = here::here("project-config.json")
)
}
\arguments{
\item{x}{date to convert}

\item{type}{character, whether to define weeks as for the "target"
(default, i.e. week definition of the forecasts/targets) or "forecast",
the forecast submission cycle, i.e. the week will end on the week
day on which forecasts are submitted.}

\item{config_file}{Path to the config file}
}
\value{
converted date
}
\description{
The day on which the week ends is given by the config variable
\code{week_end_day}
}
\author{
Sebastian Funk
}
